Automatically add ssh-keys to ssh-agent
I got tired of typing my ssh passphrases multiple times a day, so I
decided to finally do something about it.

The `ssh()` wrapper function has been updated to check, for every
`id_<name>` private key file in ~/.ssh, if there is an executable
`id_<name>.sh` file. If so, it uses it as `SSH_ASKPASS`.

In turn, there is a `id_rsa.sh` script in `~/.ssh`, that checks if there
is a file in the `pass` password store, with the path
`$(personal/ssh/$(basename $0))`, and if so decrypts it using pass and
writes it to stdout. Since it determines the password file using `$0`, I
can just make symlinks to it to support other ssh keys/passphrases.